feat(ivy): support i18n without closure
So far using runtime i18n with ivy meant that you needed to use Closure and `goog.getMsg` (or a polyfill). This PR changes the compiler to only output closure code when the compiler option `i18nUseClosure` is set to true. This option is set to false by the compiler-cli when the option `i18nInFile` is used, which is what happens when you load a translations file.
This means that if you use the Angular CLI with ivy and load a translations file, you can use i18n and the application will not throw at runtime.
For now it will not translate your application, but at least you can try ivy without having to remove all of your i18n code and configuration.